+++ tests/src/application/embeddings/embedding_generator_test.py
@@ -1,6 +1,9 @@
 from pathlib import Path
 from unittest.mock import patch
 
+import requests
+import requests_mock
+
 from src.application.embeddings.embedding_generator import EmbeddingGenerator
 
 
@@ -10,15 +13,18 @@ def test_generate(app_context):
     with open(file_path, 'r', encoding='utf-8') as f:
         html_content = f.read()
 
-    with patch('requests.api.get') as mock_requests_get, \
-        patch("langchain_experimental.text_splitter.SemanticChunker.split_text") as mock_split_text, \
+    # Initializing mock
+    session = requests.Session()
+    adapter = requests_mock.Adapter()
+
+    session.mount('http://', adapter)
+    adapter.register_uri('GET', 'http://example.com', text=html_content)
+
+    with patch("langchain_experimental.text_splitter.SemanticChunker.split_text") as mock_split_text, \
         patch('langchain_community.vectorstores.mongodb_atlas.MongoDBAtlasVectorSearch.add_documents') as mock_add_documents:
-
-        embedding_generator = EmbeddingGenerator(app_context)
-
-        mock_requests_get.return_value.text = html_content
         mock_split_text.return_value = ["chunk1", "chunk2"]
 
+        embedding_generator = EmbeddingGenerator(app_context)
         embedding_generator.generate("http://example.com")
 
         mock_split_text.assert_called_once()